[MOD/#42] TabRow 컴포넌트 파일 분리
